requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
Flask-Login==0.6.2
pg8000==1.29.3
python-dotenv==1.0.0
Werkzeug==2.3.7
reportlab==4.0.4
Jinja2==3.1.2
pandas==2.0.3
openpyxl==3.1.2
pdfkit==1.0.0
playwright==1.40.0
passlib==1.7.4
cachetools==5.3.2
flask-mail==0.9.1
json2pdf-Converter==0.5
